--- .gitattributes
@@ -1,2 +1,0 @@
-*.cpp linguist-language=cpp
-*.h linguist-language=cpp
