+++ prerefactor/logging/sqlite3.h
