+++ prerefactor/logging/sqlite3ext.h
